+++ src/main/java/uk/gov/cshr/report/controller/CourseCompletionsController.java
@@ -8,6 +8,10 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import uk.gov.cshr.report.controller.model.*;
+import uk.gov.cshr.report.controller.model.reportRequest.AddReportRequestResponse;
+import uk.gov.cshr.report.controller.model.reportRequest.GetReportRequestParams;
+import uk.gov.cshr.report.controller.model.reportRequest.GetReportRequestsResponse;
+import uk.gov.cshr.report.controller.model.reportRequest.PostCourseCompletionsReportRequestParams;
 import uk.gov.cshr.report.domain.aggregation.Aggregation;
 import uk.gov.cshr.report.domain.aggregation.CourseCompletionAggregation;
 import uk.gov.cshr.report.domain.aggregation.CourseCompletionByOrganisationAggregation;
