+++ src/main/java/uk/gov/cshr/report/controller/RegisteredLearnersController.java
@@ -5,7 +5,11 @@ import jakarta.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
-import uk.gov.cshr.report.controller.model.*;
+import uk.gov.cshr.report.controller.model.UpdateUserDetailsParams;
+import uk.gov.cshr.report.controller.model.UpdateUserResult;
+import uk.gov.cshr.report.controller.model.reportRequest.AddReportRequestResponse;
+import uk.gov.cshr.report.controller.model.reportRequest.GetReportRequestParams;
+import uk.gov.cshr.report.controller.model.reportRequest.GetReportRequestsResponse;
 import uk.gov.cshr.report.controller.model.reportRequest.OrganisationalReportRequestParams;
 import uk.gov.cshr.report.domain.report.RegisteredLearnerReportRequest;
 import uk.gov.cshr.report.service.RegisteredLearnersService;
