+++ src/main/java/uk/gov/cshr/report/service/reportRequests/CourseCompletionReportRequestService.java
@@ -3,7 +3,7 @@ package uk.gov.cshr.report.service.reportRequests;
 import org.springframework.stereotype.Service;
 import uk.gov.cshr.report.config.reports.CourseCompletionsReportConfig;
 import uk.gov.cshr.report.controller.mappers.PostCourseCompletionsReportRequestsParamsToReportRequestMapper;
-import uk.gov.cshr.report.controller.model.PostCourseCompletionsReportRequestParams;
+import uk.gov.cshr.report.controller.model.reportRequest.PostCourseCompletionsReportRequestParams;
 import uk.gov.cshr.report.domain.report.CourseCompletionReportRequest;
 import uk.gov.cshr.report.repository.ReportRequestRepository;
