+++ src/main/java/uk/gov/cshr/report/service/reportRequests/ReportRequestService.java
@@ -3,8 +3,8 @@ package uk.gov.cshr.report.service.reportRequests;
 import lombok.extern.slf4j.Slf4j;
 import uk.gov.cshr.report.config.reports.ReportExportConfig;
 import uk.gov.cshr.report.controller.mappers.ReportParamsToRequestMapper;
-import uk.gov.cshr.report.controller.model.AddReportRequestResponse;
-import uk.gov.cshr.report.controller.model.GetReportRequestParams;
+import uk.gov.cshr.report.controller.model.reportRequest.AddReportRequestResponse;
+import uk.gov.cshr.report.controller.model.reportRequest.GetReportRequestParams;
 import uk.gov.cshr.report.controller.model.reportRequest.ReportRequestParams;
 import uk.gov.cshr.report.domain.report.ReportRequest;
 import uk.gov.cshr.report.domain.report.ReportRequestStatus;
